+++ src/NCase.Core/Back.Api/CombinationSet/IndentationException.cs
@@ -2,7 +2,7 @@ using JetBrains.Annotations;
 using NDsl.Back.Api.Ex;
 using NDsl.Back.Api.Util;
 
-namespace NCaseFramework.Back.Api.Combinations
+namespace NCaseFramework.Back.Api.CombinationSet
 {
     public class IndentationException : InvalidSyntaxException
     {
