+++ src/NCase.Core/Back.Api/CombinationSet/PairwiseProdId.cs
@@ -1,6 +1,6 @@
 using JetBrains.Annotations;
 
-namespace NCaseFramework.Back.Api.Combinations
+namespace NCaseFramework.Back.Api.CombinationSet
 {
     public class PairwiseProdId : ProdId
     {
